+++ .ci/linux/install-gcc.sh
@@ -1,8 +1,8 @@
 #!/bin/sh
 
-set -e
+set -ex
 
-if [ -z ${GCC_VERSION+x} ]; then
+if [ -z "${GCC_VERSION}" ]; then
   echo "GCC_VERSION is empty. skip install gcc..."
   return
 fi
@@ -19,5 +19,10 @@ fi
 sudo apt-get -y --allow-unauthenticated update
 sudo apt-get -qq install "g++-${GCC_VERSION}"
 sudo update-alternatives --install /usr/bin/gcc gcc \
-  /usr/bin/gcc-${GCC_VERSION} 90 \
-  --slave /usr/bin/g++ g++ /usr/bin/g++-${GCC_VERSION}
+  "/usr/bin/gcc-${GCC_VERSION}" 1990 \
+  --slave /usr/bin/g++ g++ "/usr/bin/g++-${GCC_VERSION}"
+
+update-alternatives --display gcc
+
+"g++-${GCC_VERSION}" --version
+g++ --version