+++ include/internal/iutest_genparams.hpp
@@ -527,6 +527,10 @@ class iuCartesianProductHolder
     typedef tuples::tuple<const Generator...> _MyTuple;
 
 public:
+    iuCartesianProductHolder(const iuCartesianProductHolder& rhs)
+        : v(rhs.v)
+    {
+    }
     explicit iuCartesianProductHolder(const Generator&... generators)
         : v(generators...) {}
 
@@ -1093,6 +1097,8 @@ class iuPairwiseHolder
     typedef tuples::tuple<const Generator...> _MyTuple;
 
 public:
+    iuPairwiseHolder(const iuPairwiseHolder& rhs)
+        : v(rhs.v) {}
     explicit iuPairwiseHolder(const Generator&... generators)
         : v(generators...) {}