+++ include/internal/iutest_type_traits.hpp
@@ -164,6 +164,7 @@ using ::std::is_class;
 using ::std::is_convertible;
 using ::std::is_base_of;
 using ::std::is_signed;
+using ::std::is_unsigned;
 using ::std::add_lvalue_reference;
 #if IUTEST_HAS_RVALUE_REFS
 using ::std::add_rvalue_reference;
@@ -700,6 +701,12 @@ struct is_signed
 {
 };
 
+template<typename T>
+struct is_unsigned
+    : public bool_constant<!is_signed<T>::value>
+{
+};
+
 #if !defined(IUTEST_NO_TEMPLATE_PARTIAL_SPECIALIZATION)
 
 /**
@@ -970,7 +977,14 @@ using namespace has_equal_to_operator_helper;   // NOLINT
 template<typename T>
 struct has_equal_to_operator
 {
+#if IUTEST_HAS_DECLTYPE
+    typedef bool_constant< (sizeof(::std::declval<T>() == ::std::declval<T>()) != sizeof(has_equal_to_operator_helper::no_t) ) > type;    // NOLINT
+#else
+IUTEST_PRAGMA_WARN_PUSH()
+IUTEST_PRAGMA_WARN_DISABLE_NONNULL()
     typedef bool_constant< (sizeof(*(T*)0 == *(T*)0) != sizeof(has_equal_to_operator_helper::no_t) ) > type;    // NOLINT
+IUTEST_PRAGMA_WARN_POP()
+#endif
 };
 
 IUTEST_PRAGMA_WARN_POP()