+++ test/no_failure_fail_tests.cpp
@@ -60,9 +60,9 @@ int main(int argc, char* argv[])
     IUTEST_ASSERT_EQ(2, ::iutest::UnitTest::GetInstance()->failed_test_count()) << ::iutest::AssertionReturn(1);
     IUTEST_ASSERT_EQ(0, nCount) << ::iutest::AssertionReturn(1);
 #else
-    IUTEST_ASSERT_EXIT( ::iutest::UnitTest::GetInstance()->successful_test_count() == 0 );
-    IUTEST_ASSERT_EXIT( ::iutest::UnitTest::GetInstance()->failed_test_count() == 2 );
-    IUTEST_ASSERT_EXIT( nCount == 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::UnitTest::GetInstance()->successful_test_count() == 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::UnitTest::GetInstance()->failed_test_count() == 2 );
+    IUTEST_TERMINATE_ON_FAILURE( nCount == 0 );
 #endif
     printf("*** Successful ***\n");
     return 0;