+++ test/printers_tests.cpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2012-2020, Takazumi Shirayanagi\n
+ * Copyright (C) 2012-2021, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -32,8 +32,9 @@
     (void)(expect); \
     (void)(val)
 
-// unused
-// #define IUTEST_PRINTTOSTRING_CONTAIN(expect, val)
+#define IUTEST_PRINTTOSTRING_CONTAIN(expect, val)   \
+    (void)(expect); \
+    (void)(val)
 
 #endif
 
@@ -89,6 +90,58 @@ IUTEST(PrintToTest, Bar)
 }
 
 #if !defined(IUTEST_USE_GTEST)
+
+IUTEST(PrintToTest, FloatingPoint)
+{
+    ::iutest::floating_point<float> f = 1.0f;
+    ::iutest::internal::FloatingPoint<float> F(1.0f);
+    IUTEST_ASSERT_STREQ(::iutest::PrintToString(f), ::iutest::PrintToString(F));
+}
+
+#if IUTEST_HAS_INT128
+IUTEST(PrintToTest, Int128)
+{
+    typedef ::iutest::internal::TypeWithSize<16>::Int i128_t;
+    IUTEST_ASSUME_EQ(16u, sizeof(i128_t));
+    i128_t i128 = ::iutest::detail::numeric_min< ::iutest::internal::TypeWithSize<8>::Int >(); // -9223372036854775808
+    i128 -= 1;
+    LogChecker ck("0xFFFFFFFFFFFFFFFF7FFFFFFFFFFFFFFF");
+    IUTEST_PRINTTOSTRING_EQ(ck, i128);
+    IUTEST_STREAMOUT_CHECK(i128);
+}
+
+IUTEST(PrintToTest, UInt128)
+{
+    typedef ::iutest::internal::TypeWithSize<16>::UInt i128_t;
+    IUTEST_ASSUME_EQ(16u, sizeof(i128_t));
+    i128_t i128 = ::iutest::detail::numeric_max< ::iutest::internal::TypeWithSize<8>::UInt >(); // 18446744073709551615;
+    i128 += 1;
+    LogChecker ck("0x00000000000000010000000000000000");
+    IUTEST_PRINTTOSTRING_EQ(ck, i128);
+    IUTEST_STREAMOUT_CHECK(i128);
+}
+#endif
+
+#if IUTEST_HAS_LONG_DOUBLE
+IUTEST(PrintToTest, LongDouble)
+{
+    ::iutest::internal::LongDouble ld(static_cast< ::iutest::internal::LongDouble::Float>(1.0f));
+    LogChecker ck("1");
+    IUTEST_PRINTTOSTRING_CONTAIN(ck, ld);
+    IUTEST_STREAMOUT_CHECK(ld);
+}
+#endif
+
+#if IUTEST_HAS_FLOAT128
+IUTEST(PrintToTest, Float128)
+{
+    ::iutest::internal::Float128 f128(static_cast< ::iutest::internal::Float128::Float>(1.0f));
+    LogChecker ck("1");
+    IUTEST_PRINTTOSTRING_CONTAIN(ck, f128);
+    IUTEST_STREAMOUT_CHECK(f128);
+}
+#endif
+
 
 IUTEST(PrintToTest, IutestAnyNotInitialized)
 {
@@ -194,7 +247,7 @@ IUTEST(PrintToTest, String)
         IUTEST_STREAMOUT_CHECK(c);
     }
     {
-        LogChecker ck("10");
+        LogChecker ck("0x0A");
         char c = '\n';
         IUTEST_PRINTTOSTRING_EQ(ck, c);
         IUTEST_STREAMOUT_CHECK(c);
@@ -202,6 +255,13 @@ IUTEST(PrintToTest, String)
     {
         LogChecker ck("\'A\'");
         char c = 'A';
+        IUTEST_PRINTTOSTRING_EQ(ck, c);
+        IUTEST_STREAMOUT_CHECK(c);
+    }
+    {
+        LogChecker ck("0x80");
+        unsigned char uc = 0x80;
+        char c = static_cast<char>(uc);
         IUTEST_PRINTTOSTRING_EQ(ck, c);
         IUTEST_STREAMOUT_CHECK(c);
     }
@@ -239,9 +299,9 @@ IUTEST(PrintToTest, WideString)
         IUTEST_STREAMOUT_CHECK(c);
     }
     {
-        LogChecker ck("10");
+        LogChecker ck("000A");
         wchar_t c = L'\n';
-        IUTEST_PRINTTOSTRING_EQ(ck, c);
+        IUTEST_PRINTTOSTRING_CONTAIN(ck, c);
         IUTEST_STREAMOUT_CHECK(c);
     }
     {
@@ -265,7 +325,7 @@ IUTEST(PrintToTest, SurrogatePair)
 {
 #if !defined(IUTEST_USE_GTEST)
     {
-        const wchar_t* p = L"\U00020BB7野家";
+        const wchar_t* p = L"\U00020BB7";
         const ::std::string s = ::iutest::PrintToString(p);
         if( s[0] == '0' )
         {
@@ -280,7 +340,7 @@ IUTEST(PrintToTest, SurrogatePair)
         }
         else
         {
-            LogChecker ck("\U00020BB7野家");
+            LogChecker ck("\U00020BB7");
             IUTEST_PRINTTOSTRING_EQ(ck, s);
             IUTEST_STREAMOUT_CHECK(p);
         }
@@ -288,7 +348,7 @@ IUTEST(PrintToTest, SurrogatePair)
 #endif
 #if IUTEST_HAS_CHAR16_T_PRINTABLE
     {
-        const char16_t* p = u"\U00020BB7野家";
+        const char16_t* p = u"\U00020BB7";
         const ::std::string s = ::iutest::PrintToString(p);
         if( s[0] == '?' )
         {
@@ -296,7 +356,7 @@ IUTEST(PrintToTest, SurrogatePair)
         }
         else
         {
-            LogChecker ck("\U00020BB7野家");
+            LogChecker ck("\U00020BB7");
             IUTEST_PRINTTOSTRING_EQ(ck, s);
             IUTEST_STREAMOUT_CHECK(p);
         }