+++ test/set_up_testcase_failure_tests.cpp
@@ -1,12 +1,12 @@
 ﻿//======================================================================
 //-----------------------------------------------------------------------
 /**
- * @file        set_up_TestSuite_failure_tests.cpp
+ * @file        set_up_testcase_failure_tests.cpp
  * @brief       SetUpTestSuite で失敗したときのテスト
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2013-2021, Takazumi Shirayanagi\n
+ * Copyright (C) 2013-2022, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -66,20 +66,20 @@ int main(int argc, char* argv[])
     // "Google Test" fails to set up a test case,
     // it does not report a failure, and the test is also continued
     // fixed at https://github.com/google/googletest/commit/9ed99c6c837ae1cbfcabd36959fc802ebb5ae07f
-    IUTEST_ASSERT_EXIT( ret == 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ret == 0 );
     if( ret != 0 ) return 1;
 #else
-    IUTEST_ASSERT_EXIT( ret != 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ret != 0 );
     if( ret == 0 ) return 1;
 #endif
 
-    IUTEST_ASSERT_EXIT( ::iutest::UnitTest::GetInstance()->failed_test_count() != 2 );
-#if !defined(IUTEST_USE_GTEST)
-    IUTEST_ASSERT_EXIT( setup_flag == 0 );
+    IUTEST_TERMINATE_ON_FAILURE( ::iutest::UnitTest::GetInstance()->failed_test_count() != 2 );
+#if defined(IUTEST_USE_GTEST) && GTEST_VER <=  0x01110000 && !GTEST_LATEST
+    IUTEST_TERMINATE_ON_FAILURE( setup_flag == 1 );
 #else
-    IUTEST_ASSERT_EXIT( setup_flag == 1 );
+    IUTEST_TERMINATE_ON_FAILURE( setup_flag == 0 );
 #endif
-    IUTEST_ASSERT_EXIT( teardown_flag == 1);
+    IUTEST_TERMINATE_ON_FAILURE( teardown_flag == 1);
 
     printf("*** Successful ***\n");
     return 0;