+++ test/syntax_tests.cpp
@@ -6,7 +6,7 @@
  *
  * @author      t.shirayanagi
  * @par         copyright
- * Copyright (C) 2013-2020, Takazumi Shirayanagi\n
+ * Copyright (C) 2013-2021, Takazumi Shirayanagi\n
  * This software is released under the new BSD License,
  * see LICENSE
 */
@@ -17,6 +17,7 @@
 // include
 #include "iutest.hpp"
 #include "pred_tests.hpp"
+#include <limits>
 
 namespace syntax_tests
 {
@@ -245,6 +246,35 @@ IUTEST(SyntaxTest, Double)
         IUTEST_ASSUME_DOUBLE_EQ(1.0, d) << d;
 }
 
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(SyntaxTest, LongDouble)
+{
+    if( long double d = 1.0l )
+        IUTEST_ASSERT_LONG_DOUBLE_EQ(1.0l, d) << d;
+    if( long double d = 1.0l )
+        IUTEST_EXPECT_LONG_DOUBLE_EQ(1.0l, d) << d;
+    if( long double d = 1.0l )
+        IUTEST_INFORM_LONG_DOUBLE_EQ(1.0l, d) << d;
+    if( long double d = 1.0l )
+        IUTEST_ASSUME_LONG_DOUBLE_EQ(1.0l, d) << d;
+}
+
+#endif
+
+IUTEST(SyntaxTest, Complex)
+{
+    ::std::complex<float> c(1.0f, 1.0f);
+    if( float d = 1.0 )
+        IUTEST_ASSERT_COMPLEX_EQ(::std::complex<float>(d, d), c) << c;
+    if( float d = 1.0 )
+        IUTEST_EXPECT_COMPLEX_EQ(::std::complex<float>(d, d), c) << c;
+    if( float d = 1.0 )
+        IUTEST_INFORM_COMPLEX_EQ(::std::complex<float>(d, d), c) << c;
+    if( float d = 1.0 )
+        IUTEST_ASSUME_COMPLEX_EQ(::std::complex<float>(d, d), c) << c;
+}
+
 IUTEST(SyntaxTest, Near)
 {
     if( int x = 1 )
@@ -291,6 +321,35 @@ IUTEST(SyntaxTest, AlmostDoubleEq)
         IUTEST_INFORM_ALMOST_EQ(1.0, x);
     if( double x = 1.0 )
         IUTEST_ASSUME_ALMOST_EQ(1.0, x);
+}
+
+#if IUTEST_HAS_LONG_DOUBLE
+
+IUTEST(SyntaxTest, AlmostLongDoubleEq)
+{
+    if( long double x = 1.0l )
+        IUTEST_ASSERT_ALMOST_EQ(1.0l, x);
+    if( long double x = 1.0l )
+        IUTEST_EXPECT_ALMOST_EQ(1.0l, x);
+    if( long double x = 1.0l )
+        IUTEST_INFORM_ALMOST_EQ(1.0l, x);
+    if( long double x = 1.0l )
+        IUTEST_ASSUME_ALMOST_EQ(1.0l, x);
+}
+
+#endif
+
+IUTEST(SyntaxTest, AlmostComplexEq)
+{
+    ::std::complex<double> c(1.0, 0.0);
+    if( double x = 1.0 )
+        IUTEST_ASSERT_ALMOST_EQ(c, x);
+    if( double x = 1.0 )
+        IUTEST_EXPECT_ALMOST_EQ(x, c);
+    if( double x = 1.0 )
+        IUTEST_INFORM_ALMOST_EQ(::std::complex<double>(x, x), c);
+    if( double x = 1.0 )
+        IUTEST_ASSUME_ALMOST_EQ(c, ::std::complex<double>(x, x));
 }
 
 IUTEST(SyntaxTest, Null)
@@ -423,6 +482,20 @@ IUTEST(SyntaxTest, StrLn)
         IUTEST_INFORM_STRLNEQ(len, test) << len;
     if( int len = 4 )
         IUTEST_ASSUME_STRLNEQ(len, test) << len;
+}
+
+IUTEST(SyntaxTest, Nan)
+{
+    float  fnan = ::std::numeric_limits<float>::quiet_NaN();
+    double dnan = ::std::numeric_limits<double>::quiet_NaN();
+    if( fnan )
+        IUTEST_ASSERT_NAN(fnan) << fnan;
+    if( fnan )
+        IUTEST_EXPECT_NAN(fnan) << fnan;
+    if( dnan )
+        IUTEST_INFORM_NAN(dnan) << dnan;
+    if( dnan )
+        IUTEST_ASSUME_NAN(dnan) << dnan;
 }
 
 #if defined(IUTEST_OS_WINDOWS)
@@ -957,7 +1030,7 @@ IUTEST(PrintToTest, Iomanip)
 #if IUTEST_HAS_CHAR16_T
 IUTEST(PrintToTest, U16String)
 {
-    IUTEST_SUCCEED() << u"テスト";
+    IUTEST_SUCCEED() << u"test";
 }
 
 #if IUTEST_HAS_CXX_HDR_STRING_VIEW
@@ -973,7 +1046,7 @@ IUTEST(PrintToTest, U16StringStringView)
 #if IUTEST_HAS_CHAR32_T
 IUTEST(PrintToTest, U32String)
 {
-    IUTEST_SUCCEED() << U"テスト";
+    IUTEST_SUCCEED() << U"TEST";
 }
 
 #if IUTEST_HAS_CXX_HDR_STRING_VIEW