+++ test/type_param_register_exception_int_tests.cpp
@@ -1,0 +1,20 @@
+﻿//======================================================================
+//-----------------------------------------------------------------------
+/**
+ * @file        type_param_register_exception_int_tests.cpp
+ * @brief       type parameter exception in register test
+ *
+ * @author      t.shirayanagi
+ * @par         copyright
+ * Copyright (C) 2021, Takazumi Shirayanagi\n
+ * This software is released under the new BSD License,
+ * see LICENSE
+*/
+//-----------------------------------------------------------------------
+//======================================================================
+
+#include "type_param_register_exception_tests_decl.hpp"
+#if REGISTER_EXCEPTION_TEST
+#  define REGISTER_EXCEPTION_TEST_THROW_INT   1
+#endif
+#include "type_param_register_exception_tests_decl.cpp"